Folder for each algorithm used
