+++ test/inverseProblem/particleSwarm_LOG/OSTRICH.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# set path to OSTRICH binary
+OSTRICH=/home/edwin/Apps/Ostrich_v17.12.19/Linux/openmpi/2.0.2/Ostrich
+
+$OSTRICH
+
+# pause for user input
+read -n1 -r -p "Press any key to continue..." key
+
