--- test/inverseProblem/rxn_database.dat
@@ -1,7 +1,0 @@
-'temperature points' 3 0. 25. 60.
-'Vaq' 1.0 0. 100.
-'null'
-'null'
-'null'
-'null'
-'null'
