--- test/inverseProblem/templatePFLO.tpl
@@ -1,262 +1,0 @@
-#==============================================================================
-#Description: Toy problem for richards equation
-#Flow: RICHARDS
-#==============================================================================
-
-############# General Configuration ###############
-
-SIMULATION
-  SIMULATION_TYPE SUBSURFACE
-  PROCESS_MODELS
-    SUBSURFACE_FLOW flow
-      MODE RICHARDS
-    /
-   SUBSURFACE_TRANSPORT transport
-     MODE GIRT
-    /
-  /
-END
-
-############# FLOW CONFIGURATION ###############
-SUBSURFACE
-
-############# GRID ###############
-GRID
-  TYPE STRUCTURED CARTESIAN
-  NXYZ 1 1 100
-  BOUNDS
-    0.d0   0.d0   0.d0
-    0.05d0 0.05d0 0.50d0
-  /
-END
-
-############# REGIONS ###############
-REGION Region__all
-  COORDINATES
-    0.d0   0.d0   0.d0
-    0.05d0 0.05d0 0.50d0
-  /
-END
-
-REGION Face__Top
-  COORDINATES
-    0.05d0 0.05d0 0.50d0
-    0.d0 0.d0 0.50d0
-  /
-  FACE TOP
-END
-
-REGION Face__Bottom
-  COORDINATES
-    0.d0 0.d0 0.d0
-    0.05d0 0.05d0 0.d0
-  /
-  FACE BOTTOM
-END
-
-REGION Obs__PointOutflow
-  COORDINATES
-    0.025d0 0.025d0 0.50d0
-  /
-END
-
-############# OBSERVATION ###############
-OBSERVATION
-  REGION Obs__PointOutflow
-END
-
-############# FLUID ###############
-FLUID_PROPERTY
-  DIFFUSION_COEFFICIENT 1.d-9
-END
-
-############# SOILS ###############
-MATERIAL_PROPERTY soil1
-  ID 1
-  POROSITY <porosity>
-  TORTUOSITY 0.5d0
-  PERMEABILITY
-    PERM_X 1.d-4
-    PERM_Y 1.d-4
-    PERM_Z 1.d-4
-  /
-  SATURATION_FUNCTION vanGenuchtenMualem1
-  LONGITUDINAL_DISPERSIVITY 1.333452E-05
-END
-
-############# CURVES ###############
-CHARACTERISTIC_CURVES vanGenuchtenMualem1
-  SATURATION_FUNCTION VAN_GENUCHTEN
-    LIQUID_RESIDUAL_SATURATION 0.115
-    ALPHA  1.9401d-4
-    M 0.286d0 
-  /
-
-  PERMEABILITY_FUNCTION MUALEM_VG_LIQ
-    LIQUID_RESIDUAL_SATURATION 0.115
-    M 0.286d0
-  /
-END
-
-############# FLOW CONDITIONS ###############
-FLOW_CONDITION bc__inflow
- TYPE
-   FLUX NEUMANN
- END
- FLUX <darcyVel> cm/h
-END
-
-FLOW_CONDITION bc__outflow
- TYPE
-   PRESSURE DIRICHLET
- END
-  PRESSURE 101325.d0
-END
-
-FLOW_CONDITION bc__noFlow
-  TYPE
-    FLUX neumann
-  END
-  FLUX 0.d0
-END
-
-FLOW_CONDITION ic__initial
-  TYPE
-    FLUX HYDROSTATIC
-  END
-  INTERPOLATION LINEAR
-  DATUM 0.d0 0.d0 0.5d0
-  PRESSURE 101325.d0
-END
-
-########### TRANSPORT CONDITIONS #############
-TRANSPORT_CONDITION ic__initial_t
-  TYPE DIRICHLET_ZERO_GRADIENT
-  CONSTRAINT_LIST
-    0.d0 initial_constraint
-  /
-END
-
-TRANSPORT_CONDITION bc__inflow_t
-  TYPE DIRICHLET_ZERO_GRADIENT
-  CONSTRAINT_LIST
-# Pulse injection
-    0.0d0 inlet_constraint
-    <elutionTime> clean_inlet
-## Continous injection
-#   0.0d0 clean_inlet
-#   <elutionTime> inlet_constraint
-  /
-END
-
-TRANSPORT_CONDITION bc__outflow_t
-  TYPE ZERO_GRADIENT
-  CONSTRAINT_LIST
-    0.d0 initial_constraint
-  /
-END
-
-############## CONSTRAINTS #################
-CONSTRAINT initial_constraint
-  CONCENTRATIONS
-    Vaq   1.d-50      T
-  /
-  IMMOBILE
-    Vim   1.d-50
-  /
-END
-
-CONSTRAINT inlet_constraint
-  CONCENTRATIONS
-    Vaq   1.66d-16      T
-  /
-END
-
-CONSTRAINT clean_inlet
-  CONCENTRATIONS
-    Vaq   1.d-50      T
-  /
-END
-
-############# COUPLERS ###############
-# initial condition
-INITIAL_CONDITION
-  REGION Region__all
-  FLOW_CONDITION ic__initial
-  TRANSPORT_CONDITION ic__initial_t
-END
-
-# boundary conditions
-BOUNDARY_CONDITION
-  REGION Face__Bottom
-  FLOW_CONDITION bc__inflow
-  TRANSPORT_CONDITION bc__inflow_t
-END
-
-BOUNDARY_CONDITION
-  REGION Face__Top
-  FLOW_CONDITION bc__outflow
-  TRANSPORT_CONDITION bc__outflow_t
-END
-
-############# STRATAS ###############
-STRATA
-  REGION Region__all
-  MATERIAL soil1
-END
-
-############# OUTPUT ###############
-OUTPUT
-  SNAPSHOT_FILE
-  	PERIODIC TIME 2.d-1 d
-  /
-  
-  OBSERVATION_FILE
-    PERIODIC TIME 1.d-1 d
-  /
-  FORMAT HDF5
-  VELOCITY_AT_CENTER
-/
-
-############# TIME ###############
-TIME
-  FINAL_TIME <endTime> d
-  INITIAL_TIMESTEP_SIZE 1.d-2 d
-  MAXIMUM_TIMESTEP_SIZE 5.d-1 d
-/
-
-############# CHEMISTRY MAGIC ###############
-CHEMISTRY
-  PRIMARY_SPECIES
-    Vaq
-  /
-  
-  IMMOBILE_SPECIES
-    Vim
-  /
-
-  REACTION_SANDBOX
-    BIOPARTICLE
-      PARTICLE_NAME_AQ Vaq
-      PARTICLE_NAME_IM Vim
-      RATE_ATTACHMENT 9.864485E-08 1/h
-      RATE_DETACHMENT 2.494554E-05 1/h
-      DECAY_AQUEOUS 2.501787E-02 1/h
-      DECAY_ADSORBED 1.307897E-05 1/h
-      ZERO_CONCENTRATION 0.0
-    /
-  /
-  LOG_FORMULATION
-  TRUNCATE_CONCENTRATION 1.0E-50
-  DATABASE rxn_database.dat
-  OUTPUT
-    TOTAL
-    ALL
-  /
-END
-
-############# REFERENCE VALUES ###############
-
-REFERENCE_PRESSURE 101325.
-
-END_SUBSURFACE
